+++ crates/dirt_granular/src/granular_temp.rs
@@ -64,10 +64,10 @@ pub fn print_granular_temperature(
     let mut local_mv_z = 0.0;
     let mut local_mass = 0.0;
     for i in 0..nlocal {
-        local_mv_x += atoms.mass[i] * atoms.vel[i][0];
-        local_mv_y += atoms.mass[i] * atoms.vel[i][1];
-        local_mv_z += atoms.mass[i] * atoms.vel[i][2];
-        local_mass += atoms.mass[i];
+        local_mv_x += atoms.mass[i] as f64 * atoms.vel[i][0] as f64;
+        local_mv_y += atoms.mass[i] as f64 * atoms.vel[i][1] as f64;
+        local_mv_z += atoms.mass[i] as f64 * atoms.vel[i][2] as f64;
+        local_mass += atoms.mass[i] as f64;
     }
     let global_mv_x = comm.all_reduce_sum_f64(local_mv_x);
     let global_mv_y = comm.all_reduce_sum_f64(local_mv_y);
@@ -78,9 +78,9 @@ pub fn print_granular_temperature(
     let avg_vz = global_mv_z / global_mass;
     let mut vel_diff = 0.0;
     for i in 0..nlocal {
-        vel_diff += atoms.mass[i] * (atoms.vel[i][0] - avg_vx).powi(2)
-            + atoms.mass[i] * (atoms.vel[i][1] - avg_vy).powi(2)
-            + atoms.mass[i] * (atoms.vel[i][2] - avg_vz).powi(2);
+        vel_diff += atoms.mass[i] as f64 * (atoms.vel[i][0] as f64 - avg_vx).powi(2)
+            + atoms.mass[i] as f64 * (atoms.vel[i][1] as f64 - avg_vy).powi(2)
+            + atoms.mass[i] as f64 * (atoms.vel[i][2] as f64 - avg_vz).powi(2);
     }
     let vel_diff_sum = comm.all_reduce_sum_f64(vel_diff);
     let granular_temperature = vel_diff_sum / (3.0 * global_mass);
@@ -89,8 +89,8 @@ pub fn print_granular_temperature(
     let mut local_ke = 0.0;
     for i in 0..nlocal {
         local_ke += 0.5
-            * atoms.mass[i]
-            * (atoms.vel[i][0].powi(2) + atoms.vel[i][1].powi(2) + atoms.vel[i][2].powi(2));
+            * atoms.mass[i] as f64
+            * ((atoms.vel[i][0] as f64).powi(2) + (atoms.vel[i][1] as f64).powi(2) + (atoms.vel[i][2] as f64).powi(2));
     }
     let global_ke = comm.all_reduce_sum_f64(local_ke);
